hallucination-detector/requirements.txt
transformers>=4.40
bitsandbytes>=0.43
accelerate>=0.29
datasets>=2.18
scikit-learn>=1.4
numpy
pandas
scipy
pyyaml
tqdm
rich
